kafka-oracle-pg/requirements.txt
kafka-python
cx_Oracle
psycopg2